+++ sdk/communication/Azure.Communication.CallAutomation/CHANGELOG.md
@@ -5,8 +5,9 @@
 - StopDialogRequest has been removed as a parameter and object for StopDialog calls
 - Dialog calls have been separated into their own client
 - StartDialog call is now a PUT call
-- botId within StartDialog options has been renamed to botAppId
+- botId within StartDialog options has been renamed to botAppId and is now a string
 - DialogOptions is a public object now to allow it to be returned as part of the response from StartDialog
+- DialogResult now contains a DialogId
 
 ## 1.0.1-alpha.4 (Unreleased)
 - Fixed StopDialog bug